Keep native MoE unit tests on CPU and split coverage by source
Add CPU checks for native expert configuration and router-loss gradients. Move the CUDA grouped-MM comparison to an opt-in integration file and select either the functional or legacy PyTorch grouped-MM callable.

Separate utility and PP tests by their production modules while preserving existing assertions. Keep Transformers mocks tied to the module under test when real model imports replace the lazy root module.

Validation: 122 CPU tests passed. H100 numerical forward/backward comparisons passed through both grouped-MM entry points. Black, isort, and git diff --check passed.
